+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart/CartAddLineItemAction.java
@@ -5,6 +5,7 @@ import java.time.*;
 import java.time.ZonedDateTime;
 import java.util.*;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
 import javax.validation.Valid;
@@ -134,6 +135,14 @@ public interface CartAddLineItemAction
     public ExternalTaxRateDraft getExternalTaxRate();
 
     /**
+     *  <p>Sets the external Tax Rates for individual Shipping Methods, if the Cart has the <code>External</code> TaxMode and <code>Multiple</code> ShippingMode.</p>
+     * @return perMethodExternalTaxRate
+     */
+    @Valid
+    @JsonProperty("perMethodExternalTaxRate")
+    public List<MethodExternalTaxRateDraft> getPerMethodExternalTaxRate();
+
+    /**
      *  <p>Inventory mode specific to the Line Item only, and valid for the entire <code>quantity</code> of the Line Item. Set only if the inventory mode should be different from the <code>inventoryMode</code> specified on the Cart.</p>
      * @return inventoryMode
      */
@@ -239,6 +248,21 @@ public interface CartAddLineItemAction
     public void setExternalTaxRate(final ExternalTaxRateDraft externalTaxRate);
 
     /**
+     *  <p>Sets the external Tax Rates for individual Shipping Methods, if the Cart has the <code>External</code> TaxMode and <code>Multiple</code> ShippingMode.</p>
+     * @param perMethodExternalTaxRate values to be set
+     */
+
+    @JsonIgnore
+    public void setPerMethodExternalTaxRate(final MethodExternalTaxRateDraft... perMethodExternalTaxRate);
+
+    /**
+     *  <p>Sets the external Tax Rates for individual Shipping Methods, if the Cart has the <code>External</code> TaxMode and <code>Multiple</code> ShippingMode.</p>
+     * @param perMethodExternalTaxRate values to be set
+     */
+
+    public void setPerMethodExternalTaxRate(final List<MethodExternalTaxRateDraft> perMethodExternalTaxRate);
+
+    /**
      *  <p>Inventory mode specific to the Line Item only, and valid for the entire <code>quantity</code> of the Line Item. Set only if the inventory mode should be different from the <code>inventoryMode</code> specified on the Cart.</p>
      * @param inventoryMode value to be set
      */
@@ -285,6 +309,7 @@ public interface CartAddLineItemAction
         instance.setExternalPrice(template.getExternalPrice());
         instance.setExternalTotalPrice(template.getExternalTotalPrice());
         instance.setExternalTaxRate(template.getExternalTaxRate());
+        instance.setPerMethodExternalTaxRate(template.getPerMethodExternalTaxRate());
         instance.setInventoryMode(template.getInventoryMode());
         instance.setShippingDetails(template.getShippingDetails());
         instance.setCustom(template.getCustom());
@@ -317,6 +342,11 @@ public interface CartAddLineItemAction
             com.commercetools.api.models.cart.ExternalLineItemTotalPrice.deepCopy(template.getExternalTotalPrice()));
         instance.setExternalTaxRate(
             com.commercetools.api.models.cart.ExternalTaxRateDraft.deepCopy(template.getExternalTaxRate()));
+        instance.setPerMethodExternalTaxRate(Optional.ofNullable(template.getPerMethodExternalTaxRate())
+                .map(t -> t.stream()
+                        .map(com.commercetools.api.models.cart.MethodExternalTaxRateDraft::deepCopy)
+                        .collect(Collectors.toList()))
+                .orElse(null));
         instance.setInventoryMode(template.getInventoryMode());
         instance.setShippingDetails(
             com.commercetools.api.models.cart.ItemShippingDetailsDraft.deepCopy(template.getShippingDetails()));