+++ changes.md
@@ -3,6 +3,7 @@
 <details>
 <summary>Added Property(s)</summary>
 
+- added property `perMethodExternalTaxRate` to type `CartAddLineItemAction`
 - added property `shippingMode` to type `MyCartDraft`
 </details>